refactor: re-organize file & folder structure
